refactored Comment component and moved buttons to CommentButtons component
